Update Quiz Master
I am never touching this game after this. The code is some of the worst I have ever seen.

I fixed the problems that kept me up all night for the past 2 years. The sprites are fixed, the UI scales with screen size, the score calculates correctly, the progress bar is correct, there is menu navigation.